+++ sv-parser-parser/src/source_text/package_items.rs
@@ -80,8 +80,7 @@ pub(crate) fn package_or_generate_item_declaration(
 pub(crate) fn anonymous_program(s: Span) -> IResult<Span, AnonymousProgram> {
     let (s, a) = keyword("program")(s)?;
     let (s, b) = symbol(";")(s)?;
-    let (s, c) = many0(anonymous_program_item)(s)?;
-    let (s, d) = keyword("endprogram")(s)?;
+    let (s, (c, d)) = many_till(anonymous_program_item, keyword("endprogram"))(s)?;
     Ok((
         s,
         AnonymousProgram {