+++ sv-parser-parser/src/udp_declaration_and_instantiation/udp_ports.rs
@@ -45,8 +45,7 @@ pub(crate) fn udp_output_declaration(s: Span) -> IResult<Span, UdpOutputDeclarat
 #[tracable_parser]
 #[packrat_parser]
 pub(crate) fn udp_output_declaration_nonreg(s: Span) -> IResult<Span, UdpOutputDeclaration> {
-    let (s, a) = many0(attribute_instance)(s)?;
-    let (s, b) = keyword("output")(s)?;
+    let (s, (a, b)) = many_till(attribute_instance, keyword("output"))(s)?;
     let (s, c) = port_identifier(s)?;
     Ok((
         s,
@@ -57,8 +56,7 @@ pub(crate) fn udp_output_declaration_nonreg(s: Span) -> IResult<Span, UdpOutputD
 #[tracable_parser]
 #[packrat_parser]
 pub(crate) fn udp_output_declaration_reg(s: Span) -> IResult<Span, UdpOutputDeclaration> {
-    let (s, a) = many0(attribute_instance)(s)?;
-    let (s, b) = keyword("output")(s)?;
+    let (s, (a, b)) = many_till(attribute_instance, keyword("output"))(s)?;
     let (s, c) = keyword("reg")(s)?;
     let (s, d) = port_identifier(s)?;
     let (s, e) = opt(pair(symbol("="), constant_expression))(s)?;
@@ -73,8 +71,7 @@ pub(crate) fn udp_output_declaration_reg(s: Span) -> IResult<Span, UdpOutputDecl
 #[tracable_parser]
 #[packrat_parser]
 pub(crate) fn udp_input_declaration(s: Span) -> IResult<Span, UdpInputDeclaration> {
-    let (s, a) = many0(attribute_instance)(s)?;
-    let (s, b) = keyword("input")(s)?;
+    let (s, (a, b)) = many_till(attribute_instance, keyword("input"))(s)?;
     let (s, c) = list_of_udp_port_identifiers(s)?;
     Ok((s, UdpInputDeclaration { nodes: (a, b, c) }))
 }
@@ -82,8 +79,7 @@ pub(crate) fn udp_input_declaration(s: Span) -> IResult<Span, UdpInputDeclaratio
 #[tracable_parser]
 #[packrat_parser]
 pub(crate) fn udp_reg_declaration(s: Span) -> IResult<Span, UdpRegDeclaration> {
-    let (s, a) = many0(attribute_instance)(s)?;
-    let (s, b) = keyword("reg")(s)?;
+    let (s, (a, b)) = many_till(attribute_instance, keyword("reg"))(s)?;
     let (s, c) = variable_identifier(s)?;
     Ok((s, UdpRegDeclaration { nodes: (a, b, c) }))
 }